MAINT: stats._xp_mean: move _xp_mean
